refactor(api): extract token store into internal/api/tokenstore leaf package
First step of decomposing the ~13k-LOC internal/api monolith (ADR-0006).
token_store.go imported nothing from api, so it moves cleanly to a sibling
leaf package: TokenStore, ScopedToken, TokenScope (+Scope{ReadOnly,ReadWrite,
Admin}), NewTokenStore, LoadTokenFile, RequiredScopeForMethod, and the Err*
values now live in internal/api/tokenstore.

- Server.tokens and internal/daemon's token seeding import the leaf directly
  (tokenstore is now a reusable building block, not buried in transport).
- Tests of unexported internals (buildSnapshot, parseScope, tokenFileSchema)
  move with the package as package tokenstore; the one Server-level test
  (public-paths /__version) stays in package api via the exported surface.
- depguard gains api-tokenstore-isolated: the leaf may not import
  internal/api, so the boundary is CI-enforced (modelled on
  domain-core-inward-only). ADR-0006 + index.

No behavioural change: the route registry, /__capabilities, and middleware
composition order are untouched. Verified: go build, go vet, golangci-lint
(0 issues), go test -race ./internal/api/... ./internal/daemon/... (all pass),
gofumpt clean, route-policy gate green.
