+++ docs/adr/README.md
@@ -11,5 +11,6 @@ people and the diffs. Format matches the sibling repos (seed/stem).
 | [0003](0003-dependency-direction-depguard.md) | Dependency direction enforced by depguard | Accepted |
 | [0004](0004-scope-based-auth-posture.md) | Scope-based bearer-token auth + CSRF posture | Accepted |
 | [0005](0005-ed25519-signed-licenses.md) | Ed25519-signed license tokens | Accepted |
+| [0006](0006-internal-api-sub-package-decomposition.md) | Decompose `internal/api` into isolated sub-packages | Accepted |
 
 Status values: Proposed · Accepted · Amended · Superseded.
