+++ internal/api/tokenstore/store.go
@@ -1,4 +1,4 @@
-package api
+package tokenstore
 
 import (
 	"crypto/sha256"
